backend/requirements.txt
asgiref==3.7.2
awscli==1.29.44
boto3==1.28.44
botocore==1.31.44
colorama==0.4.4
Django==4.1.7
django-cors-headers==3.14.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docutils==0.16
jmespath==1.0.1
numpy==2.2.1
packaging==23.1
pandas==2.2.3
phonenumbers==8.13.19
Pillow==10.0.0
psycopg2-binary==2.9.7
py-bcrypt==0.4
pyasn1==0.5.0
PyJWT==2.8.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.6.2
setuptools==75.8.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
